refactor(calculationCard): split modals into components
