+++ src/components/calculos/calculationCard/calculationCard.jsx
@@ -1,12 +1,11 @@
-import React, { useEffect, useState } from 'react'
-import { Calculator, ChartPie, FileChartColumnIncreasing, FileDown, FileSpreadsheet, RefreshCcw, SquareActivity } from 'lucide-react';
-import ExcelViewer from './excelViewer';
-import { DistribuicaoPage } from '../../../routes/distribuicao/page';
+import React, { useEffect, useState } from 'react';
+import { Calculator, FileChartColumnIncreasing, FileDown, RefreshCcw, SquareActivity } from 'lucide-react';
 import { Atom } from 'react-loading-indicators';
-import { da, se } from 'date-fns/locale';
 import dayjs from 'dayjs';
-import domtoimage from 'dom-to-image'; 
 import { useInterval } from 'react-use';
+import SimulacaoModal from './simulacaoModal/simulacaoModal';
+import HistoricoModal from './historicoModal/historicoModal';
+import DistribuicoesModal from './distribuicoesModal/distribuicoesModal';
 
 export const CalculationCard = ({categoria, dataFiltro}) => {
     const [Ufs, setUFs] = useState(['SP', 'GO', 'MG', 'MS', 'MT', 'PA', 'RO', 'TO', 'BA']);
@@ -15,16 +14,12 @@ export const CalculationCard = ({categoria, dataFiltro}) => {
     const [resultados, setResultados] = useState({});
     const [calcular, setCalcular] = useState (false);
     const [erro, setErro] = useState(false);
-    const [mensagemCalculo, setMensagemCalculo] = useState(false);
     const [mensagemPersistir, setMensagemPersistir] = useState(false);
     const [loading, setLoading] = useState(false);
     const [outlier, setOutlier] = useState(true);
     const [distribuicoes, setDistribuicoes] = useState(false);
     const [colapsos, setColapsos] = useState({});
-    const [viewDistribuicoes, setViewDistribuicoes] = useState('p');
     const [simulacao, setSimulacao] = useState([]);
-    const distribuicoesRef = React.useRef();
-    const [fullscreenDistribuicoes, setFullscreenDistribuicoes] = useState(false);
 
     const fetchData = async (categoria, dataFiltro) => {
         try {
@@ -61,33 +56,6 @@ export const CalculationCard = ({categoria, dataFiltro}) => {
         }
     }
 
-    const baixarImagemDistribuicoes = async () => {
-        if (!distribuicoesRef.current) return;
-
-        try {
-            // 1. Expandir modal
-            setFullscreenDistribuicoes(true);
-
-            // 2. Espera o layout aplicar
-            await new Promise(requestAnimationFrame);
-
-            // 3. Gera imagem
-            const dataUrl = await domtoimage.toPng(distribuicoesRef.current);
-
-            // 4. Baixa a imagem
-            const link = document.createElement('a');
-            link.download = 'distribuicoes.png';
-            link.href = dataUrl;
-            link.click();
-        } catch (error) {
-            console.error('Erro ao gerar imagem:', error);
-        } finally {
-            // 5. Volta ao estado normal
-            setFullscreenDistribuicoes(false);
-        }
-        };
-
-
     useInterval(() => {
         fetchData(categoria, dataFiltro);
         console.log('5min');
@@ -101,7 +69,7 @@ export const CalculationCard = ({categoria, dataFiltro}) => {
     const historicModal = () => setIsModalOpen(true);
 
     const calcularModal = () => dataFiltro == dayjs().format('YYYY-MM-DD') ? setCalcular(true) : setCalcular(false);
-    const closeModalCalcular = () => { setCalcular(false); setErro(false); setMensagemCalculo(false); setOutlier(true); setSimulacao([]); setMensagemPersistir(false); };
+    const closeModalCalcular = () => { setCalcular(false); setErro(false); setOutlier(true); setSimulacao([]); setMensagemPersistir(false); };
 
     const closeModal = () => setIsModalOpen(false);
     const closeMadalDistribuicoes = () => setDistribuicoes(false);
@@ -123,242 +91,35 @@ export const CalculationCard = ({categoria, dataFiltro}) => {
 
     return (
         <>
-            {calcular && dataFiltro == dayjs().format('YYYY-MM-DD') && (
-                <div className="fixed inset-0 bg-black bg-opacity-50 flex justify-center items-center z-50">
-                    <div className=" bg-card-bg p-6 rounded-lg shadow-lg w-[80%]">
-                        <div className="flex justify-between items-center w-full mb-4">
-                            <h2 className="text-xl font-bold text-white">Simulação de Cálculo</h2>
-                            <button className="px-4 py-2 bg-button-hover-bg text-black font-bold rounded-full hover:bg-slate-400" onClick={closeModalCalcular}>X</button>
-                        </div>
-                        <div className='flex flex-wrap md:flex-nowrap space-x-2 mb-6 w-full md:w-1/2'>
-                            <span className='font-bold'> {UFSelecionada} {outlier ? 'com outliers' : 'sem outliers'}  </span>
-                            <button
-                                className={`px-4 rounded-full font-bold ${outlier ? 'bg-button-hover-bg text-black' : 'bg-[#F5F5DC] text-black hover:bg-button-hover-bg hover:text-black'}`}
-                                onClick={() => setOutlier(!outlier)}>
-                                {outlier ? 'Ativado' : 'Desativado'}
-                            </button>
-                        </div>
-                        <p className="text-base text-white mb-4">Deseja recalcular os dados para <span className="font-bold"> {UFSelecionada} </span> - <span className="font-bold">{categoria.toUpperCase()}</span> {outlier ? 'COM Outliers' : 'SEM Outliers'} ?</p>
-                        <button
-                            className={`${loading ? 'hidden' : ''} px-4 py-2 bg-text-bold-color text-black font-bold rounded-full hover:bg-slate-400`}
-                            onClick={async () => {
-                                try {
-                                    setSimulacao([]);
-                                    setMensagemCalculo(false);
-                                    setLoading(true);
-                                    setMensagemPersistir(false);
+            <SimulacaoModal
+                show={calcular && dataFiltro == dayjs().format("YYYY-MM-DD")}
+                onClose={closeModalCalcular}
+                UFSelecionada={UFSelecionada}
+                outlier={outlier}
+                setOutlier={setOutlier}
+                categoria={categoria}
+                simulacao={simulacao}
+                setSimulacao={setSimulacao}
+                mensagemPersistir={mensagemPersistir}
+                setMensagemPersistir={setMensagemPersistir}
+                setLoading={setLoading}
+                loading={loading}
+                setErro={setErro}
+                erro={erro}
+            />
 
-                                    const response = await fetch (
-                                        `https://pecuaria.datagro.com/backoffice-pec/api/v1/dashboard/calcular?obter=calcular&outlier=${outlier === true ? 1 : 0}&uf=${UFSelecionada}&resumo=1`,
-                                        { method: 'GET', headers: { 'Content-Type': 'application/json' } }
-                                    );
-                                    setLoading(false);
-                                    if (!response.ok) setErro(true);
-                                    const result = await response.json();                                   
-                                    setSimulacao(result || []);
+            <HistoricoModal
+                show={isModalOpen}
+                onClose={closeModal}
+                log={resultados[UFSelecionada]?.log}
+            />
 
-                                    if (!erro) setMensagemCalculo(true);
-                                } catch (error) {
-                                    setLoading(false);
-                                    console.error('Erro ao calcular:', error);
-                                }
-                            }}
-                        >
-                            {loading ? '' : 'Simular'}
-                        </button>
-                        
-                        {simulacao?.output && (
-                            <a
-                                className="px-4 py-2 text-button-hover-bg font-bold rounded-full hover:cursor-pointer hover:underline mt-2 "
-                                onClick={async () => {
-                                        setSimulacao([]);
-                                        setMensagemCalculo(false);
-                                        setLoading(true);
-                                        
-                                        const response = await fetch (
-                                        `https://pecuaria.datagro.com/backoffice-pec/api/v1/dashboard/calcular?obter=calcular&outlier=${outlier === true ? 1 : 0}&uf=${UFSelecionada}&resumo=0`,
-                                        { method: 'GET', headers: { 'Content-Type': 'application/json' } });
-
-                                        if (!response.ok) setErro(true);
-                                        setLoading(false);
-                                        setMensagemPersistir(true);
-                                    }}
-                             >
-                                Persistir cálculo
-                            </a>
-                        )}
-
-                    { mensagemPersistir && (
-                        <div className="text-button-hover-bg mt-4">
-                            <p>ATENÇÃO: VALOR ATUALIZADO!</p>
-                        </div>
-                    )}
-
-                        {loading && (
-                            <div className="flex justify-center items-center mt-4">
-                                <Atom color="#FFD700" size="medium" text="calculando..." textColor="" />
-                            </div>
-                        )}
-
-                        {erro && (
-                            <div className="text-red-500 mt-4">
-                                <p>Erro ao simular. Tente novamente mais tarde.</p>
-                            </div>
-                        )}
-
-                        {simulacao?.output && (
-                        <div className="mt-4 px-2">
-                            <h3 className="text-lg font-bold text-white mb-4">Simulação</h3>
-
-                            <div className="overflow-x-auto pb-2">
-                            <div
-                                className="
-                                flex space-x-4
-                                w-full
-                                pb-2
-                                "
-                            >
-                                {Object.entries(simulacao.output).flatMap(([uf, codigos]) =>
-                                Object.entries(codigos).map(([codigo, item]) => (
-                                    <div
-                                    key={codigo}
-                                    className="
-                                        bg-[#1f2937]
-                                        rounded-lg
-                                        px-4 py-3
-                                        shadow
-                                        hover:shadow-md
-                                        transition
-                                        duration-300
-                                        min-w-[220px]
-                                        text-left
-                                        flex-shrink-0
-                                    "
-                                    >
-                                    <div className="border-b border-gray-600 pb-1 mb-2">
-                                        <h4 className="text-white text-base font-semibold truncate">{codigo}</h4>
-                                        <div className="flex justify-between items-center gap-1 mt-3">
-                                            <div className="flex gap-1">
-                                                <p className="text-xs">{item.display === "b" ? "Bônus "+ (item.categoria).toUpperCase() : (item.categoria).toUpperCase()},</p>
-                                                <p className="text-xs">{(uf).toUpperCase()}</p>
-                                                <p className="text-xs">{(item.regiao || "").toUpperCase()}</p>
-                                            </div>
-                                            <p className="text-xs">{item.data}</p>
-                                            </div>
-                                    </div>
-                                    
-                                    <div className='flex flex-col items-center mb-2'>
-                                        <p className="text-white text-xl font-semibold">{item.indicador?.toFixed(2) ?? 'N/A'}</p>
-                                    </div>
-
-                                        <div className="text-gray-300 text-sm space-y-1 break-words">
-                                        <div className="flex justify-between">
-                                            <span className="font-semibold">Min:</span>
-                                            <span>{item.min?.toFixed(4) ?? 'N/A'}</span>
-                                        </div>
-                                        <div className="flex justify-between">
-                                            <span className="font-semibold">Max:</span>
-                                            <span>{item.max?.toFixed(4) ?? 'N/A'}</span>
-                                        </div>
-                                        <div className="flex justify-between">
-                                            <span className="font-semibold">Desvio Padrão:</span>
-                                            <span>{item.desvioPadrao?.toFixed(4) ?? 'N/A'}</span>
-                                        </div>
-                                        <div className="flex justify-between">
-                                            <span className="font-semibold">Outliers:</span>
-                                            <span>{item.outliers ?? 'N/A'}</span>
-                                        </div>
-                                        <div className="flex justify-between">
-                                            <span className="font-semibold">Cap:</span>
-                                            <span className="text-right">
-                                            {item.cap == null
-                                                ? 'N/A'
-                                                : Array.isArray(item.cap)
-                                                ? (item.cap.length ? item.cap.join(', ') : '[]')
-                                                : typeof item.cap === 'object'
-                                                    ? Object.entries(item.cap).map(([key, val]) => `${key}: ${(val).toFixed(4)}`).join(', ')
-                                                    : item.cap}
-                                            </span>
-                                        </div>
-                                        </div>
-
-                                    </div>
-                                ))
-                                )}
-                            </div>
-                            </div>
-                        </div>
-                        )}
-                    </div>
-                </div>
-            )}
-
-            {isModalOpen && (
-                <div className="fixed inset-0 bg-black bg-opacity-50 flex justify-center items-center z-50">
-                    <div className="bg-card-bg  p-6 rounded-lg shadow-lg w-[50%]">
-                        <div className="flex justify-between items-center w-full mb-4">
-                            <h2 className="text-xl font-bold dark:text-white">Histórico detalhado</h2>
-                            <button className="px-4 py-2 bg-button-hover-bg text-black font-bold rounded-full hover:bg-slate-400" onClick={closeModal}>X</button>
-                        </div>
-                        <div className="overflow-y-auto h-96">
-                            {resultados[UFSelecionada]?.log?.split('\n').map((line, idx) => (
-                                <p key={idx} className="text-base text-gray-300 my-1">{line}</p>
-                            ))}
-                        </div>
-                    </div>
-                </div>
-            )}
-
-            {distribuicoes && (
-            <div className="fixed inset-0 bg-black bg-opacity-50 flex justify-center items-center z-50">
-                <div
-                className={`
-                    bg-card-bg p-6 rounded-lg shadow-lg relative transition-all duration-300
-                    ${fullscreenDistribuicoes ? 'h-[200vh]' : 'w-[80%] h-[85%]'}
-                `}
-                >
-                <div className="flex gap-4 mb-2">
-                    Visualizar
-                    <button onClick={() => setViewDistribuicoes('p')}>
-                    <FileSpreadsheet className={`text-${viewDistribuicoes === 'p' ? 'text-bold-color' : 'white'}`} />
-                    </button>
-                    <button onClick={() => setViewDistribuicoes('g')}>
-                    <ChartPie className={`text-${viewDistribuicoes === 'g' ? 'text-bold-color' : 'white'}`} />
-                    </button>
-                </div>
-                <div className="flex justify-between items-center w-full mb-4">
-                    <h2 className="text-xl font-bold dark:text-white">Distribuições</h2>
-                    <div className="flex gap-2">
-                    <button
-                        className="px-4 py-2 bg-button-hover-bg text-black font-bold rounded-full hover:bg-slate-400"
-                        onClick={baixarImagemDistribuicoes}
-                    >
-                        Baixar imagem
-                    </button>
-                    <button
-                        className="px-4 py-2 bg-button-hover-bg text-black font-bold rounded-full hover:bg-slate-400"
-                        onClick={closeMadalDistribuicoes}
-                    >
-                        X
-                    </button>
-                    </div>
-                </div>
-                <div
-                    ref={distribuicoesRef}
-                    className={`
-                    justify-center items-center mb-4 w-full
-                    ${fullscreenDistribuicoes ? '' : 'overflow-y-auto h-[85%]'}
-                    `}
-                >
-                    {viewDistribuicoes === 'p' ? (
-                    <ExcelViewer url={`https://pecuaria.datagro.com/backoffice-pec/api/v1/dashboard/arquivos?acao=distribuicao&categoria=${categoria}&data=${dataFiltro}`} size={fullscreenDistribuicoes}/>
-                    ) : (
-                    <DistribuicaoPage dataFiltro={dataFiltro} />
-                    )}
-                </div>
-                </div>
-            </div>
-            )}
+            <DistribuicoesModal
+                show={distribuicoes}
+                onClose={closeMadalDistribuicoes}
+                categoria={categoria}
+                dataFiltro={dataFiltro}
+            />
 
             <div className='flex flex-col  mt-6 bg-gradient-to-br from-[#2A3529] to-[#1A1F1A] relative '> 
                 <div className='flex flex-wrap items-center justify-between px-4 '>
@@ -414,7 +175,6 @@ export const CalculationCard = ({categoria, dataFiltro}) => {
                         <button title='Atualizar' onClick={async () => {
 
                             setSimulacao([]);
-                            setMensagemCalculo(false);
                             setLoading(true);
 
                             const response = await fetch(
